requirements.txt
websocket-client
paho-mqtt
python-dotenv
